+++ gaztec/src/components/home/home.tsx
@@ -1,13 +1,29 @@
 import { css } from "@emotion/react";
 import { ContractComponent } from "../contract/contract";
 import { SidebarComponent } from "../sidebar/sidebar";
-import { useState } from "react";
-import { AztecContext } from "../../aztecEnv";
+import { useEffect, useState } from "react";
+import { AztecContext, AztecEnv } from "../../aztecEnv";
+import NoSleep from "nosleep.js";
+import { LogPanel } from "../logPanel/logPanel";
+import logoURL from "../../assets/Aztec_logo.png";
+import { CircularProgress, Drawer, LinearProgress } from "@mui/material";
 
 const layout = css({
   display: "flex",
   flexDirection: "row",
   height: "100%",
+});
+
+const logo = css({
+  width: "100%",
+  padding: "0.5rem",
+});
+
+const collapsedDrawer = css({
+  height: "100%",
+  width: "4rem",
+  backgroundColor: "var(--mui-palette-primary-light)",
+  overflow: "hidden",
 });
 
 export default function Home() {
@@ -21,6 +37,20 @@ export default function Home() {
   const [currentContract, setCurrentContract] = useState(null);
   const [currentTx, setCurrentTx] = useState(null);
   const [currentContractAddress, setCurrentContractAddress] = useState(null);
+  const [logs, setLogs] = useState([]);
+  const [logsOpen, setLogsOpen] = useState(false);
+  const [drawerOpen, setDrawerOpen] = useState(false);
+
+  const [isNetworkStoreInitialized, setIsNetworkStoreInitialized] =
+    useState(false);
+
+  useEffect(() => {
+    const initNetworkStore = async () => {
+      await AztecEnv.initNetworkStore();
+      setIsNetworkStoreInitialized(true);
+    };
+    initNetworkStore();
+  }, []);
 
   const AztecContextInitialValue = {
     pxe,
@@ -33,6 +63,12 @@ export default function Home() {
     currentTx,
     node,
     currentContractAddress,
+    logs,
+    logsOpen,
+    drawerOpen,
+    setDrawerOpen,
+    setLogsOpen,
+    setLogs,
     setAztecNode,
     setCurrentTx,
     setWalletDB,
@@ -48,7 +84,30 @@ export default function Home() {
   return (
     <div css={layout}>
       <AztecContext.Provider value={AztecContextInitialValue}>
-        <SidebarComponent />
+        <div css={collapsedDrawer} onClick={() => setDrawerOpen(!drawerOpen)}>
+          <img css={logo} src={logoURL} />
+        </div>
+        <Drawer
+          sx={{
+            "& .MuiDrawer-paper": {
+              height: "100%",
+              width: "340px",
+            },
+          }}
+          ModalProps={{
+            keepMounted: true,
+          }}
+          onClose={() => setDrawerOpen(false)}
+          variant="temporary"
+          open={drawerOpen}
+        >
+          {isNetworkStoreInitialized ? (
+            <SidebarComponent />
+          ) : (
+            <LinearProgress />
+          )}
+        </Drawer>
+        <LogPanel />
         <ContractComponent />
       </AztecContext.Provider>
     </div>