+++ l1-contracts/test/Rollup.t.sol
@@ -33,8 +33,8 @@ import {
 } from "@aztec/core/libraries/RollupLibs/ProposeLib.sol";
 
 import {
-  Timestamp, Slot, Epoch, SlotLib, EpochLib, TimeFns
-} from "@aztec/core/libraries/TimeMath.sol";
+  Timestamp, Slot, Epoch, SlotLib, EpochLib, TimeLib
+} from "@aztec/core/libraries/TimeLib.sol";
 
 // solhint-disable comprehensive-interface
 
@@ -42,10 +42,13 @@ import {
  * Blocks are generated using the `integration_l1_publisher.test.ts` tests.
  * Main use of these test is shorter cycles when updating the decoder contract.
  */
-contract RollupTest is DecoderBase, TimeFns {
+contract RollupTest is DecoderBase {
   using SlotLib for Slot;
   using EpochLib for Epoch;
   using ProposeLib for ProposeArgs;
+  using TimeLib for Timestamp;
+  using TimeLib for Slot;
+  using TimeLib for Epoch;
 
   Registry internal registry;
   Inbox internal inbox;
@@ -65,7 +68,16 @@ contract RollupTest is DecoderBase, TimeFns {
   uint256 internal privateKey;
   address internal signer;
 
-  constructor() TimeFns(TestConstants.AZTEC_SLOT_DURATION, TestConstants.AZTEC_EPOCH_DURATION) {}
+  uint256 internal SLOT_DURATION;
+  uint256 internal EPOCH_DURATION;
+
+  constructor() {
+    TimeLib.initialize(
+      block.timestamp, TestConstants.AZTEC_SLOT_DURATION, TestConstants.AZTEC_EPOCH_DURATION
+    );
+    SLOT_DURATION = TestConstants.AZTEC_SLOT_DURATION;
+    EPOCH_DURATION = TestConstants.AZTEC_EPOCH_DURATION;
+  }
 
   /**
    * @notice  Set up the contracts needed for the tests with time aligned to the provided block name
@@ -306,7 +318,7 @@ contract RollupTest is DecoderBase, TimeFns {
     rollup.propose(args, signatures, data.body, data.blobInputs);
 
     quote.epochToProve = Epoch.wrap(1);
-    quote.validUntilSlot = toSlots(Epoch.wrap(2));
+    quote.validUntilSlot = Epoch.wrap(2).toSlots();
     signedQuote = _quoteToSignedQuote(quote);
     rollup.claimEpochProofRight(signedQuote);
     BlockLog memory blockLog = rollup.getBlock(0);
@@ -334,7 +346,7 @@ contract RollupTest is DecoderBase, TimeFns {
   function testMissingProofSlashesBond(uint256 _slotToHit) public setUpFor("mixed_block_1") {
     Slot lower = rollup.getCurrentSlot() + Slot.wrap(2 * EPOCH_DURATION);
     Slot upper = Slot.wrap(
-      (type(uint256).max - Timestamp.unwrap(rollup.GENESIS_TIME())) / rollup.SLOT_DURATION()
+      (type(uint256).max - Timestamp.unwrap(rollup.getGenesisTime())) / rollup.getSlotDuration()
     );
     Slot slotToHit = Slot.wrap(bound(_slotToHit, lower.unwrap(), upper.unwrap()));
 
@@ -351,7 +363,7 @@ contract RollupTest is DecoderBase, TimeFns {
 
   function testClaimTwice() public setUpFor("mixed_block_1") {
     _testBlock("mixed_block_1", false, 1);
-    quote.validUntilSlot = toSlots(Epoch.wrap(1e9));
+    quote.validUntilSlot = Epoch.wrap(1e9).toSlots();
     signedQuote = _quoteToSignedQuote(quote);
 
     rollup.claimEpochProofRight(signedQuote);
@@ -378,7 +390,7 @@ contract RollupTest is DecoderBase, TimeFns {
 
   function testClaimOutsideClaimPhase() public setUpFor("mixed_block_1") {
     _testBlock("mixed_block_1", false, 1);
-    quote.validUntilSlot = toSlots(Epoch.wrap(1e9));
+    quote.validUntilSlot = Epoch.wrap(1e9).toSlots();
     signedQuote = _quoteToSignedQuote(quote);
     warpToL2Slot(EPOCH_DURATION + rollup.CLAIM_DURATION_IN_L2_SLOTS());
 
@@ -395,7 +407,7 @@ contract RollupTest is DecoderBase, TimeFns {
   function testNoPruneWhenClaimExists() public setUpFor("mixed_block_1") {
     _testBlock("mixed_block_1", false, 1);
 
-    quote.validUntilSlot = toSlots(Epoch.wrap(2));
+    quote.validUntilSlot = Epoch.wrap(2).toSlots();
     signedQuote = _quoteToSignedQuote(quote);
 
     warpToL2Slot(EPOCH_DURATION + rollup.CLAIM_DURATION_IN_L2_SLOTS() - 1);
@@ -411,7 +423,7 @@ contract RollupTest is DecoderBase, TimeFns {
   function testPruneWhenClaimExpires() public setUpFor("mixed_block_1") {
     _testBlock("mixed_block_1", false, 1);
 
-    quote.validUntilSlot = toSlots(Epoch.wrap(2));
+    quote.validUntilSlot = Epoch.wrap(2).toSlots();
     signedQuote = _quoteToSignedQuote(quote);
 
     warpToL2Slot(EPOCH_DURATION + rollup.CLAIM_DURATION_IN_L2_SLOTS() - 1);
@@ -433,7 +445,7 @@ contract RollupTest is DecoderBase, TimeFns {
   function testClaimAfterPrune() public setUpFor("mixed_block_1") {
     _testBlock("mixed_block_1", false, 1);
 
-    quote.validUntilSlot = toSlots(Epoch.wrap(3));
+    quote.validUntilSlot = Epoch.wrap(3).toSlots();
     signedQuote = _quoteToSignedQuote(quote);
 
     warpToL2Slot(EPOCH_DURATION + rollup.CLAIM_DURATION_IN_L2_SLOTS() - 1);
@@ -444,14 +456,14 @@ contract RollupTest is DecoderBase, TimeFns {
 
     rollup.prune();
 
-    _testBlock("mixed_block_1", false, toSlots(Epoch.wrap(3)).unwrap());
+    _testBlock("mixed_block_1", false, Epoch.wrap(3).toSlots().unwrap());
 
     quote.epochToProve = Epoch.wrap(3);
     signedQuote = _quoteToSignedQuote(quote);
 
     vm.expectEmit(true, true, true, true);
     emit IRollup.ProofRightClaimed(
-      quote.epochToProve, quote.prover, address(this), quote.bondAmount, toSlots(Epoch.wrap(3))
+      quote.epochToProve, quote.prover, address(this), quote.bondAmount, Epoch.wrap(3).toSlots()
     );
     rollup.claimEpochProofRight(signedQuote);
   }
@@ -591,7 +603,7 @@ contract RollupTest is DecoderBase, TimeFns {
     bytes32 inboxRoot2 = inbox.getRoot(2);
 
     BlockLog memory blockLog = rollup.getBlock(1);
-    Slot prunableAt = blockLog.slotNumber + toSlots(Epoch.wrap(2));
+    Slot prunableAt = blockLog.slotNumber + Epoch.wrap(2).toSlots();
 
     Timestamp timeOfPrune = rollup.getTimestampForSlot(prunableAt);
     vm.warp(Timestamp.unwrap(timeOfPrune));
@@ -658,7 +670,7 @@ contract RollupTest is DecoderBase, TimeFns {
     _testBlock("mixed_block_1", false, 1);
     assertEq(rollup.getEpochToProve(), 0, "Invalid epoch to prove");
     warpToL2Slot(EPOCH_DURATION * 2);
-    _testBlock("mixed_block_1", false, toSlots(Epoch.wrap(2)).unwrap());
+    _testBlock("mixed_block_1", false, Epoch.wrap(2).toSlots().unwrap());
 
     assertEq(rollup.getPendingBlockNumber(), 1, "Invalid pending block number");
     assertEq(rollup.getProvenBlockNumber(), 0, "Invalid proven block number");
@@ -763,7 +775,7 @@ contract RollupTest is DecoderBase, TimeFns {
     BlockLog memory blockLog = rollup.getBlock(0);
 
     quote.epochToProve = Epoch.wrap(1);
-    quote.validUntilSlot = toSlots(Epoch.wrap(2));
+    quote.validUntilSlot = Epoch.wrap(2).toSlots();
     signedQuote = _quoteToSignedQuote(quote);
 
     warpToL2Slot(EPOCH_DURATION + rollup.CLAIM_DURATION_IN_L2_SLOTS() - 1);