+++ spartan/terraform/deploy-release/main.tf
@@ -29,6 +29,15 @@ provider "helm" {
   }
 }
 
+
+data "terraform_remote_state" "metrics" {
+  backend = "gcs"
+  config = {
+    bucket = "aztec-terraform"
+    prefix = "metrics-deploy/us-west1-a/aztec-gke-private/metrics/terraform.tfstate"
+  }
+}
+
 # Aztec Helm release for gke-cluster
 resource "helm_release" "aztec-gke-cluster" {
   provider         = helm.gke-cluster
@@ -88,10 +97,34 @@ resource "helm_release" "aztec-gke-cluster" {
   }
 
   dynamic "set" {
-    for_each = var.ETHEREUM_EXTERNAL_HOST != "" ? toset(["iterate"]) : toset([])
+    for_each = var.EXTERNAL_ETHEREUM_HOST != "" ? toset(["iterate"]) : toset([])
     content {
       name  = "ethereum.externalHost"
-      value = var.ETHEREUM_EXTERNAL_HOST
+      value = var.EXTERNAL_ETHEREUM_HOST
+    }
+  }
+
+  dynamic "set" {
+    for_each = var.EXTERNAL_ETHEREUM_CONSENSUS_HOST != "" ? toset(["iterate"]) : toset([])
+    content {
+      name  = "ethereum.beacon.externalHost"
+      value = var.EXTERNAL_ETHEREUM_CONSENSUS_HOST
+    }
+  }
+
+  dynamic "set" {
+    for_each = var.EXTERNAL_ETHEREUM_CONSENSUS_HOST_API_KEY != "" ? toset(["iterate"]) : toset([])
+    content {
+      name  = "ethereum.beacon.apiKey"
+      value = var.EXTERNAL_ETHEREUM_CONSENSUS_HOST_API_KEY
+    }
+  }
+
+  dynamic "set" {
+    for_each = var.EXTERNAL_ETHEREUM_CONSENSUS_HOST_API_KEY_HEADER != "" ? toset(["iterate"]) : toset([])
+    content {
+      name  = "ethereum.beacon.apiKeyHeader"
+      value = var.EXTERNAL_ETHEREUM_CONSENSUS_HOST_API_KEY_HEADER
     }
   }
 
@@ -101,8 +134,13 @@ resource "helm_release" "aztec-gke-cluster" {
   }
 
   set {
-    name  = "telemetry.useGcloudObservability"
-    value = "true"
+    name  = "telemetry.otelCollectorEndpoint"
+    value = "http://${data.terraform_remote_state.metrics.outputs.otel_collector_ip}:4318"
+  }
+
+  set {
+    name  = "network.gke"
+    value = true
   }
 
   # Setting timeout and wait conditions