+++ yarn-project/end-to-end/src/e2e_l1_with_wall_time.test.ts
@@ -1,6 +1,4 @@
-import { getSchnorrAccount } from '@aztec/accounts/schnorr';
-import { type InitialAccountData } from '@aztec/accounts/testing';
-import { FeeJuicePaymentMethod, type Logger, type PXE, TxStatus } from '@aztec/aztec.js';
+import { type Logger, type PXE, Wallet } from '@aztec/aztec.js';
 import { EthAddress } from '@aztec/circuits.js';
 import { getL1ContractsConfigEnvVars } from '@aztec/ethereum';
 import { type PXEService } from '@aztec/pxe';
@@ -9,14 +7,15 @@ import { jest } from '@jest/globals';
 import { privateKeyToAccount } from 'viem/accounts';
 
 import { getPrivateKeyFromIndex, setup } from './fixtures/utils.js';
+import { submitTxsTo } from './shared/submit-transactions.js';
 
 jest.setTimeout(1000 * 60 * 10);
 
 describe('e2e_l1_with_wall_time', () => {
   let logger: Logger;
   let teardown: () => Promise<void>;
+  let wallet: Wallet;
   let pxe: PXE;
-  let initialFundedAccounts: InitialAccountData[];
 
   const deploymentsPerBlock = 8;
   const numberOfBlocks = 4;
@@ -26,11 +25,10 @@ describe('e2e_l1_with_wall_time', () => {
     const initialValidators = [EthAddress.fromString(account.address)];
     const { ethereumSlotDuration } = getL1ContractsConfigEnvVars();
 
-    ({ teardown, logger, pxe, initialFundedAccounts } = await setup(0, {
+    ({ teardown, logger, wallet, pxe } = await setup(1, {
       initialValidators,
       ethereumSlotDuration,
       salt: 420,
-      numberOfInitialFundedAccounts: deploymentsPerBlock * numberOfBlocks,
     }));
   });
 
@@ -38,10 +36,7 @@ describe('e2e_l1_with_wall_time', () => {
 
   it('should produce blocks with a bunch of transactions', async () => {
     for (let i = 0; i < numberOfBlocks; i++) {
-      const txs = await submitTxsTo(
-        pxe as PXEService,
-        initialFundedAccounts.slice(i * deploymentsPerBlock, (i + 1) * deploymentsPerBlock),
-      );
+      const txs = await submitTxsTo(pxe as PXEService, deploymentsPerBlock, wallet, logger);
       await Promise.all(
         txs.map(async (tx, j) => {
           logger.info(`Waiting for tx ${i}-${j}: ${await tx.getTxHash()} to be mined`);
@@ -50,28 +45,4 @@ describe('e2e_l1_with_wall_time', () => {
       );
     }
   });
-
-  // submits a set of transactions to the provided Private eXecution Environment (PXE)
-  const submitTxsTo = async (pxe: PXEService, accounts: InitialAccountData[]) => {
-    return await Promise.all(
-      accounts.map(async account => {
-        const accountManager = await getSchnorrAccount(pxe, account.secret, account.signingKey, account.salt);
-        const paymentMethod = new FeeJuicePaymentMethod(account.address);
-        const tx = accountManager.deploy({ fee: { paymentMethod } });
-        const txHash = await tx.getTxHash();
-        logger.info(`Tx sent with hash ${txHash}`);
-
-        const receipt = await tx.getReceipt();
-        expect(receipt).toEqual(
-          expect.objectContaining({
-            status: TxStatus.PENDING,
-            error: '',
-          }),
-        );
-        logger.info(`Receipt received for ${txHash}`);
-
-        return tx;
-      }),
-    );
-  };
 });