+++ yarn-project/end-to-end/src/fixtures/utils.ts
@@ -821,13 +821,18 @@ function createDelayedL1TxUtils(aztecNodeConfig: AztecNodeConfig, privateKey: `0
   return l1TxUtils;
 }
 
-export async function createForwarderContract(aztecNodeConfig: AztecNodeConfig, privateKey: `0x${string}`) {
+export async function createForwarderContract(
+  aztecNodeConfig: AztecNodeConfig,
+  privateKey: `0x${string}`,
+  rollupAddress: Hex,
+) {
   const { walletClient, publicClient } = createL1Clients(aztecNodeConfig.l1RpcUrl, privateKey, foundry);
   const forwarderContract = await ForwarderContract.create(
     walletClient.account.address,
     walletClient,
     publicClient,
     createLogger('forwarder'),
+    rollupAddress,
   );
   return forwarderContract;
 }